pyproject.toml
# ╔══════════════════════════════════════════════════════════════════╗
# ║  NEXUS SPECTER PRO — pyproject.toml                             ║
# ║  Modern Python build system (PEP 517/518/621)                   ║
# ║  by OPTIMIUM NEXUS LLC | contact@optimiumnexus.com              ║
# ╚══════════════════════════════════════════════════════════════════╝

[build-system]
requires      = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

# ── Project metadata (PEP 621) ───────────────────────────────────────────
[project]
name        = "nexus-specter-pro"
version     = "1.1.0"
description = "Military-Grade Automated Offensive Penetration Testing Platform"
readme      = "README.md"
license     = { text = "Commercial — OPTIMIUM NEXUS LLC" }
authors     = [{ name = "OPTIMIUM NEXUS LLC", email = "contact@optimiumnexus.com" }]
maintainers = [{ name = "OPTIMIUM NEXUS LLC", email = "contact@optimiumnexus.com" }]

keywords = [
    "penetration-testing", "pentest", "red-team", "offensive-security",
    "cybersecurity", "vulnerability-scanner", "automation",
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX :: Linux",
    "Environment :: Console",
]

requires-python = ">=3.12"

dependencies = [
    # Core
    "rich>=13.7.0",
    "click>=8.1.7",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    # AI
    "anthropic>=0.25.0",
    # Security / crypto
    "cryptography>=42.0.0",
    "pyOpenSSL>=24.0.0",
    # OSINT
    "shodan>=1.31.0",
    "dnspython>=2.6.0",
    "python-whois>=0.9.4",
    # Reporting
    "jinja2>=3.1.3",
    "weasyprint>=62.0",
    # Dashboard
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "pydantic>=2.6.1",
    # Database
    "sqlalchemy>=2.0.25",
    "redis>=5.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.2",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=5.0.0",
    "black>=24.2.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
    "pre-commit>=3.7.0",
]
full = [
    "celery>=5.3.6",
    "impacket>=0.11.0",
    "scapy>=2.5.0",
    "pandas>=2.2.0",
    "matplotlib>=3.8.3",
    "selenium>=4.18.0",
    "paramiko>=3.4.0",
]

[project.urls]
Homepage      = "https://www.optimiumnexus.com"
Repository    = "https://github.com/optimiumnexusllc/NEXUS-SPECTER-PRO"
"Bug Tracker" = "https://github.com/optimiumnexusllc/NEXUS-SPECTER-PRO/issues"
Contact       = "mailto:contact@optimiumnexus.com"

[project.scripts]
nsp = "nsp_cli:main"

# ── setuptools config ────────────────────────────────────────────────────
[tool.setuptools]
packages = { find = { where = ["."], exclude = ["tests*", "docs*"] } }

[tool.setuptools.package-data]
"nsp.reporting.templates" = ["*.html", "*.css"]
"config"                  = ["*.yaml", "*.yml"]
"missions"                = ["*.yaml"]

# ── Black (formatter) ────────────────────────────────────────────────────
[tool.black]
line-length    = 100
target-version = ["py312"]
skip-magic-trailing-comma = true

# ── Ruff (linter) ────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501", "B008", "B904"]

[tool.ruff.isort]
known-first-party = ["nsp"]

# ── mypy (type checking) ─────────────────────────────────────────────────
[tool.mypy]
python_version         = "3.12"
ignore_missing_imports = true
warn_return_any        = true
warn_unused_configs    = true
exclude                = ["tests/", "docs/"]

# ── pytest ────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths    = ["tests"]
asyncio_mode = "auto"
addopts      = "-v --tb=short --cov=nsp --cov-report=term-missing"
filterwarnings = ["ignore::DeprecationWarning"]

# ── coverage ──────────────────────────────────────────────────────────────
[tool.coverage.run]
source   = ["nsp"]
omit     = ["tests/*", "docs/*", "*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

# ── pre-commit hooks ──────────────────────────────────────────────────────
# Run: pre-commit install
# Config in .pre-commit-config.yaml
